requirements.txt
asgiref==3.3.1
cffi==1.14.4
cryptography==3.2.1
Django==3.1.6
django-extensions==3.1.1
django-filter==2.4.0
djangorestframework==3.12.2
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
